refactor: shorten response directory
