+++ src/momento/responses/data/list/list_pop_front.py
