+++ src/momento/responses/data/list/list_remove_value.py
