+++ src/momento/responses/data/scalar/delete.py
