+++ src/momento/responses/data/scalar/get.py
